chore: clean up eol.ts, move logic into api
